+++ kinto/core/storage/memory.py
@@ -303,12 +303,11 @@ def extract_record_set(records, filters, sorting,
     filtered = list(apply_filters(records, filters or []))
     total_records = len(filtered)
 
-    paginated = {}
-    for rule in pagination_rules or []:
-        values = list(apply_filters(filtered, rule))
-        paginated.update(dict(((x[id_field], x) for x in values)))
-
-    if paginated:
+    if pagination_rules:
+        paginated = {}
+        for rule in pagination_rules:
+            values = apply_filters(filtered, rule)
+            paginated.update(((x[id_field], x) for x in values))
         paginated = paginated.values()
     else:
         paginated = filtered